+++ wxc_sdk/as_api.py
@@ -52,31 +52,31 @@ CALLING_DATA_TIMEOUT_PROTECTION = False
 
 
 __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'AsAnnouncementApi',
-           'AsAnnouncementsRepositoryApi', 'AsApiChild', 'AsAppServicesApi', 'AsApplyLineKeyTemplatesJobsApi',
-           'AsAttachmentActionsApi', 'AsAuthorizationsApi', 'AsAutoAttendantApi', 'AsAvailableNumbersApi',
-           'AsBargeApi', 'AsCQPolicyApi', 'AsCallBridgeApi', 'AsCallInterceptApi', 'AsCallParkApi', 'AsCallPickupApi',
-           'AsCallQueueApi', 'AsCallRecordingApi', 'AsCallRecordingSettingsApi', 'AsCallWaitingApi', 'AsCallerIdApi',
-           'AsCallingBehaviorApi', 'AsCallparkExtensionApi', 'AsCallsApi', 'AsDECTDevicesApi', 'AsDetailedCDRApi',
-           'AsDeviceConfigurationsApi', 'AsDeviceSettingsJobsApi', 'AsDevicesApi', 'AsDialPlanApi',
-           'AsDigitPatternsApi', 'AsDndApi', 'AsEventsApi', 'AsExecAssistantApi', 'AsForwardingApi', 'AsGroupsApi',
-           'AsGuestManagementApi', 'AsHotelingApi', 'AsHuntGroupApi', 'AsIncomingPermissionsApi',
-           'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi', 'AsLocationAccessCodesApi', 'AsLocationInterceptApi',
-           'AsLocationMoHApi', 'AsLocationNumbersApi', 'AsLocationVoicemailSettingsApi', 'AsLocationsApi',
-           'AsManageNumbersJobsApi', 'AsMeetingChatsApi', 'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi',
-           'AsMeetingParticipantsApi', 'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi',
-           'AsMeetingTranscriptsApi', 'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi',
-           'AsMusicOnHoldApi', 'AsNumbersApi', 'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi',
-           'AsOutgoingPermissionsApi', 'AsPagingApi', 'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi',
-           'AsPersonSettingsApiChild', 'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPrivacyApi',
-           'AsPrivateNetworkConnectApi', 'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi',
-           'AsReceptionistContactsDirectoryApi', 'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi',
-           'AsRoomsApi', 'AsRouteGroupApi', 'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi',
-           'AsScimApiChild', 'AsScimV2Api', 'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi',
-           'AsTelephonyDevicesApi', 'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi',
-           'AsVirtualLinesApi', 'AsVoiceMessagingApi', 'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi',
-           'AsVoicemailRulesApi', 'AsWebexSimpleApi', 'AsWebhookApi', 'AsWorkspaceDevicesApi',
-           'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi', 'AsWorkspaceNumbersApi',
-           'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi']
+           'AsAnnouncementsRepositoryApi', 'AsApiChild', 'AsAppServicesApi', 'AsAppSharedLineApi',
+           'AsApplyLineKeyTemplatesJobsApi', 'AsAttachmentActionsApi', 'AsAuthorizationsApi', 'AsAutoAttendantApi',
+           'AsAvailableNumbersApi', 'AsBargeApi', 'AsCQPolicyApi', 'AsCallBridgeApi', 'AsCallInterceptApi',
+           'AsCallParkApi', 'AsCallPickupApi', 'AsCallQueueApi', 'AsCallRecordingApi', 'AsCallRecordingSettingsApi',
+           'AsCallWaitingApi', 'AsCallerIdApi', 'AsCallingBehaviorApi', 'AsCallparkExtensionApi', 'AsCallsApi',
+           'AsDECTDevicesApi', 'AsDetailedCDRApi', 'AsDeviceConfigurationsApi', 'AsDeviceSettingsJobsApi',
+           'AsDevicesApi', 'AsDialPlanApi', 'AsDigitPatternsApi', 'AsDndApi', 'AsEventsApi', 'AsExecAssistantApi',
+           'AsForwardingApi', 'AsGroupsApi', 'AsGuestManagementApi', 'AsHotelingApi', 'AsHuntGroupApi',
+           'AsIncomingPermissionsApi', 'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi',
+           'AsLocationAccessCodesApi', 'AsLocationInterceptApi', 'AsLocationMoHApi', 'AsLocationNumbersApi',
+           'AsLocationVoicemailSettingsApi', 'AsLocationsApi', 'AsManageNumbersJobsApi', 'AsMeetingChatsApi',
+           'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
+           'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi', 'AsMeetingTranscriptsApi',
+           'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsMonitoringApi', 'AsMusicOnHoldApi', 'AsNumbersApi',
+           'AsOrganisationVoicemailSettingsAPI', 'AsOrganizationApi', 'AsOutgoingPermissionsApi', 'AsPagingApi',
+           'AsPeopleApi', 'AsPersonForwardingApi', 'AsPersonSettingsApi', 'AsPersonSettingsApiChild',
+           'AsPreferredAnswerApi', 'AsPremisePstnApi', 'AsPrivacyApi', 'AsPrivateNetworkConnectApi',
+           'AsPushToTalkApi', 'AsRebuildPhonesJobsApi', 'AsReceptionistApi', 'AsReceptionistContactsDirectoryApi',
+           'AsRecordingsApi', 'AsReportsApi', 'AsRestSession', 'AsRoomTabsApi', 'AsRoomsApi', 'AsRouteGroupApi',
+           'AsRouteListApi', 'AsSCIM2BulkApi', 'AsSCIM2UsersApi', 'AsScheduleApi', 'AsScimApiChild', 'AsScimV2Api',
+           'AsStatusAPI', 'AsTeamMembershipsApi', 'AsTeamsApi', 'AsTelephonyApi', 'AsTelephonyDevicesApi',
+           'AsTelephonyLocationApi', 'AsTransferNumbersApi', 'AsTrunkApi', 'AsVirtualLinesApi', 'AsVoiceMessagingApi',
+           'AsVoicePortalApi', 'AsVoicemailApi', 'AsVoicemailGroupsApi', 'AsVoicemailRulesApi', 'AsWebexSimpleApi',
+           'AsWebhookApi', 'AsWorkspaceDevicesApi', 'AsWorkspaceLocationApi', 'AsWorkspaceLocationFloorApi',
+           'AsWorkspaceNumbersApi', 'AsWorkspacePersonalizationApi', 'AsWorkspaceSettingsApi', 'AsWorkspacesApi']
 
 
 @dataclass(init=False)
@@ -6038,6 +6038,161 @@ class AsAppServicesApi(AsPersonSettingsApiChild):
         await self.put(ep, params=params, data=data)
 
 
+class AsAppSharedLineApi(AsApiChild, base='telephony/config/people'):
+    """
+    Webex app shared line API
+    """
+
+    def f_ep(self, person_id: str, application_id: str, path: str) -> str:
+        """
+
+        :meta private:
+        """
+        return super().ep(f'{person_id}/applications/{application_id}/{path}')
+
+    def search_members_gen(self, person_id: str, application_id: str, max_: str = None,
+                       start: str = None, location: str = None, name: str = None,
+                       number: str = None, order: str = None,
+                       extension: str = None, **params) -> AsyncGenerator[MemberCommon, None, None]:
+        """
+        Search Shared-Line Appearance Members
+
+        Get members available for shared-line assignment to a Webex Calling Apps Desktop device.
+
+        This API requires a full or user administrator or location administrator auth token with
+        the `spark-admin:people_read` scope.
+
+        :param person_id: A unique identifier for the person.
+        :type person_id: str
+        :param application_id: A unique identifier for the application.
+        :type application_id: str
+        :param max_: Number of records per page.
+        :type max_: str
+        :param start: Page number.
+        :type start: str
+        :param location: Location ID for the user.
+        :type location: str
+        :param name: Search for users with names that match the query.
+        :type name: str
+        :param number: Search for users with numbers that match the query.
+        :type number: str
+        :param order: Sort by first name (`fname`) or last name (`lname`).
+        :type order: str
+        :param extension: Search for users with extensions that match the query.
+        :type extension: str
+        :return: Generator yielding :class:`MemberCommon` instances
+        """
+        if max_ is not None:
+            params['max'] = max_
+        if start is not None:
+            params['start'] = start
+        if location is not None:
+            params['location'] = location
+        if name is not None:
+            params['name'] = name
+        if number is not None:
+            params['number'] = number
+        if order is not None:
+            params['order'] = order
+        if extension is not None:
+            params['extension'] = extension
+        url = self.f_ep(person_id=person_id, application_id=application_id, path='availableMembers')
+        return self.session.follow_pagination(url, model=MemberCommon, params=params, item_key='members')
+
+    async def search_members(self, person_id: str, application_id: str, max_: str = None,
+                       start: str = None, location: str = None, name: str = None,
+                       number: str = None, order: str = None,
+                       extension: str = None, **params) -> List[MemberCommon]:
+        """
+        Search Shared-Line Appearance Members
+
+        Get members available for shared-line assignment to a Webex Calling Apps Desktop device.
+
+        This API requires a full or user administrator or location administrator auth token with
+        the `spark-admin:people_read` scope.
+
+        :param person_id: A unique identifier for the person.
+        :type person_id: str
+        :param application_id: A unique identifier for the application.
+        :type application_id: str
+        :param max_: Number of records per page.
+        :type max_: str
+        :param start: Page number.
+        :type start: str
+        :param location: Location ID for the user.
+        :type location: str
+        :param name: Search for users with names that match the query.
+        :type name: str
+        :param number: Search for users with numbers that match the query.
+        :type number: str
+        :param order: Sort by first name (`fname`) or last name (`lname`).
+        :type order: str
+        :param extension: Search for users with extensions that match the query.
+        :type extension: str
+        :return: Generator yielding :class:`MemberCommon` instances
+        """
+        if max_ is not None:
+            params['max'] = max_
+        if start is not None:
+            params['start'] = start
+        if location is not None:
+            params['location'] = location
+        if name is not None:
+            params['name'] = name
+        if number is not None:
+            params['number'] = number
+        if order is not None:
+            params['order'] = order
+        if extension is not None:
+            params['extension'] = extension
+        url = self.f_ep(person_id=person_id, application_id=application_id, path='availableMembers')
+        return [o async for o in self.session.follow_pagination(url, model=MemberCommon, params=params, item_key='members')]
+
+    async def get_members(self, person_id: str, application_id: str) -> DeviceMembersResponse:
+        """
+        Get Shared-Line Appearance Members
+
+        Get primary and secondary members assigned to a shared line on a Webex Calling Apps Desktop device.
+
+        This API requires a full or user administrator or location administrator auth token with
+        the `spark-admin:people_read` scope.
+
+        :param person_id: A unique identifier for the person.
+        :type person_id: str
+        :param application_id: A unique identifier for the application.
+        :type application_id: str
+        :rtype: :class:`GetSharedLineMemberList`
+        """
+        url = self.f_ep(person_id=person_id, application_id=application_id, path='members')
+        data = await super().get(url)
+        r = DeviceMembersResponse.model_validate(data)
+        return r
+
+    async def update_members(self, person_id: str, application_id: str,
+                       members: list[DeviceMember] = None):
+        """
+        Put Shared-Line Appearance Members
+
+        Add or modify primary and secondary users assigned to shared-lines on a Webex Calling Apps Desktop device.
+
+        This API requires a full or user administrator or location administrator auth token with the
+        `spark-admin:people_write` scope.
+
+        :param person_id: A unique identifier for the person.
+        :type person_id: str
+        :param application_id: A unique identifier for the application.
+        :type application_id: str
+        :type members: list[DeviceMember]
+        :rtype: None
+        """
+        body = dict()
+        if members is not None:
+            body['members'] = TypeAdapter(list[DeviceMember]).dump_python(members, mode='json', by_alias=True,
+                                                                          exclude_none=True)
+        url = self.f_ep(person_id=person_id, application_id=application_id, path='members')
+        await super().put(url, json=body)
+
+
 class AsAvailableNumbersApi(AsApiChild, base='telephony/config'):
     """
     API for person's available numbers
@@ -8924,6 +9079,7 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
 
     #: agent caller id Api
     agent_caller_id: AsAgentCallerIdApi
+    app_shared_line: AsAppSharedLineApi
     #: Person's Application Services Settings
     appservices: AsAppServicesApi
     #: Available numbers for a person
@@ -8976,6 +9132,7 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
     def __init__(self, session: AsRestSession):
         super().__init__(session=session)
         self.agent_caller_id = AsAgentCallerIdApi(session=session)
+        self.app_shared_line = AsAppSharedLineApi(session=session)
         self.appservices = AsAppServicesApi(session=session)
         self.available_numbers = AsAvailableNumbersApi(session=session)
         self.barge = AsBargeApi(session=session)